+++ Sources/HeroDetail/HeroDetailView.swift
@@ -2,12 +2,12 @@ import Kingfisher
 import LeanRedux
 import SwiftUI
 
-public struct HeroDetailView: View
+struct HeroDetailView: View
 {
     // MARK: Dependencies
 
     @EnvironmentObject
-    private var store: HeroDetailReducer.Store
+    private var store: HeroDetailFeature.Store
 
     @Environment(\.lexemes)
     private var lexemes: Lexemes
@@ -19,7 +19,7 @@ public struct HeroDetailView: View
         State(state: store.state)
     }
 
-    public var body: some View
+    var body: some View
     {
         ZStack{
             HeroDetailBackgroundView()
@@ -48,15 +48,17 @@ public struct HeroDetailView: View
 
     // MARK: Lifecycle
 
-    public static func make(heroId: Int) -> some View
+    static func make(
+        heroId: Int,
+        middlewares: [HeroDetailFeature.Store.Middleware]
+    ) -> some View
     {
         HeroDetailView()
             .environmentObject(
-                HeroDetailReducer.Store.make(
-                    middlewares: [],
-                    reducer: HeroDetailReducer.reduce,
-                    state: HeroDetailReducer.State()
-                        .with(\.selectedHeroId, heroId)
+                HeroDetailFeature.Store.make(
+                    middlewares: middlewares,
+                    reducer: HeroDetailFeature.reduce,
+                    state: HeroDetailFeature.State().with(\.selectedHeroId, heroId)
                 )
             )
     }
@@ -68,7 +70,7 @@ public struct HeroDetailView: View
         var imageUrl: URL?
         var isDescriptionExpanded: Bool
 
-        init(state: HeroDetailReducer.State)
+        init(state: HeroDetailFeature.State)
         {
             self.imageUrl = state.thumbnail
             self.isDescriptionExpanded = state.isDescriptionExpanded
@@ -78,7 +80,7 @@ public struct HeroDetailView: View
 
 // MARK: - Preview
 
-let state = HeroDetailReducer.State()
+let state = HeroDetailFeature.State()
     .with(\.name, "3-D Man")
     .with(\.heroDescription, """
     The 3-D Man was a 1950's hero who came about through the unique merger of two brothers, Hal and Chuck Chandler. Chuck was a test pilot who was abducted by alien Skrulls during an important test flight. Earth was seen as a strategic location in the ongoing conflict between the alien Kree and Skrull Empires, so the Skrulls were seeking information on Earth's space program and had captured Chuck to interrogate him. Chuck resisted and escaped, accidentally causing the explosion of the Skrull spacecraft in the process. While his brother Hal watched, the radiation from the explosion seemingly disintegrated Chuck, who disappeared in a burst of light. Hal later discovered, however, that the light burst had imprinted an image of Chuck on each lens of Hal's eyeglasses. Through concentration, Hal could merge the images and cause Chuck to reappear as a three-dimensional man. Chuck become the costumed adventurer known as the 3-D Man and single-handedly subverted the Skrulls' early attempts to undermine Earthly civilization.
@@ -86,12 +88,9 @@ let state = HeroDetailReducer.State()
     .with(\.thumbnail, nil)
 
 
-struct HeroDetailView_Previews: PreviewProvider
+#Preview
 {
-    static var previews: some View
-    {
-        HeroDetailView.make(heroId: Int())
-    }
+    HeroDetailView.make(heroId: Int(), middlewares: [])
 }
 
 struct HeroDetailView_Default_Previews: PreviewProvider
@@ -101,7 +100,7 @@ struct HeroDetailView_Default_Previews: PreviewProvider
     {
         HeroDetailView()
             .environmentObject(
-                HeroDetailReducer.Store.make(
+                HeroDetailFeature.Store.make(
                     state: state
                 )
             )
@@ -116,7 +115,7 @@ struct HeroDetailView_Expanded_Previews: PreviewProvider
     {
         HeroDetailView()
             .environmentObject(
-                HeroDetailReducer.Store.make(
+                HeroDetailFeature.Store.make(
                     state: state
                         .with(\.isDescriptionExpanded, true)
                 )
@@ -132,7 +131,7 @@ struct HeroDetailView_Liked_Previews: PreviewProvider
     {
         HeroDetailView()
             .environmentObject(
-                HeroDetailReducer.Store.make(
+                HeroDetailFeature.Store.make(
                     state: state
                         .with(\.liked, true)
                 )