+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added mock generation script.
 - Routing target.
 - Navigator component.
+- Hero Detail dismiss button.
 
 ### Changed